Create ArrangeTheWords.java
We define a sentence to be a string of space-separated words that starts with a capital letter followed by lowercase letters and spaces, and ends with a period, i.e., it satisfies the regular expression ^[A-Z][a-z ]*\.$. We want to rearrange the words in a sentence such that the following conditions are satisfied:

 

Each word is ordered by length, ascending.
Words of equal length must appear in the same order as in the original sentence.
The rearranged sentence must be formatted to satisfy the regular expression ^[A-Z][a-z ]*\.$
 

For example, consider the sentence Cats and hats. First the words are ordered by length, maintaining original order for ties: [and, cats, hats]. Now reassemble the sentence, applying formatting: And cats hats.

 

Function Description 

Complete the function arrange in the editor below. The function must return a properly formed sentence arranged as described.

 

arrange has the following parameter(s):

    sentence:  a well formed sentence string

 Sample Input 0

The lines are printed in reverse order.
Sample Output 0

In the are lines order printed reverse.